+++ lib/src/elements/filter/mod.rs
@@ -36,7 +36,9 @@ mod video_synchronizer;
 pub use audio_resampler::{AudioResampler, AudioResamplerError};
 pub use audio_volume::{AudioVolume, AudioVolumeError, AudioVolumeHandle, AudioVolumeOptions};
 pub use change_gate::ChangeGate;
-pub use chroma_key::{ChromaKeyMethod, ChromaKeyOptions, SwChromaKey, SwChromaKeyError};
+pub use chroma_key::{
+    ChromaKeyHandle, ChromaKeyMethod, ChromaKeyOptions, SwChromaKey, SwChromaKeyError,
+};
 #[cfg(all(target_os = "windows", feature = "d3d11"))]
 pub use chroma_key::{D3d11ChromaKey, D3d11ChromaKeyError};
 #[cfg(feature = "cuda")]