+++ lib/src/elements/mod.rs
@@ -66,11 +66,12 @@ pub use driver::{
 };
 pub use filter::{
     AudioCodec, AudioResampler, AudioResamplerError, AudioVolume, AudioVolumeError,
-    AudioVolumeHandle, AudioVolumeOptions, ChangeGate, ChromaKeyMethod, ChromaKeyOptions,
-    FrameRateLimiter, Pacer, PacerError, PauseGate, PauseGateHandle, SwAudioEncoder,
-    SwAudioEncoderError, SwAudioEncoderOptions, SwChromaKey, SwChromaKeyError, SwDecoder,
-    SwDecoderError, SwEncoder, SwEncoderError, SwEncoderOptions, SwScaler, SwScalerError, Tee,
-    TeeBuilder, TeeHandle, TimestampOrigin, VideoCodec, VideoSynchronizer, VideoSynchronizerError,
+    AudioVolumeHandle, AudioVolumeOptions, ChangeGate, ChromaKeyHandle, ChromaKeyMethod,
+    ChromaKeyOptions, FrameRateLimiter, Pacer, PacerError, PauseGate, PauseGateHandle,
+    SwAudioEncoder, SwAudioEncoderError, SwAudioEncoderOptions, SwChromaKey, SwChromaKeyError,
+    SwDecoder, SwDecoderError, SwEncoder, SwEncoderError, SwEncoderOptions, SwScaler,
+    SwScalerError, Tee, TeeBuilder, TeeHandle, TimestampOrigin, VideoCodec, VideoSynchronizer,
+    VideoSynchronizerError,
 };
 #[cfg(feature = "cuda")]
 pub use filter::{